requirements.txt
fastapi
uvicorn
gspread
gspread-formatting
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests
